+++ crates/client/rpc/src/base/mod.rs
@@ -2,5 +2,4 @@ pub(crate) mod block;
 pub(crate) mod meter;
 pub(crate) mod meter_rpc;
 pub(crate) mod traits;
-pub(crate) mod transaction_rpc;
 pub(crate) mod types;
