+++ crates/client/txpool/src/lib.rs
@@ -9,5 +9,10 @@ pub use events::{EventLog, Pool, TxEvent};
 mod exex;
 pub use exex::tracex_exex;
 
+mod rpc;
+pub use rpc::{
+    Status, TransactionStatusApiImpl, TransactionStatusApiServer, TransactionStatusResponse,
+};
+
 mod tracker;
 pub use tracker::Tracker;
